Return i64 as row count.

+++ yt/ytlib/table_client/channel_writer.cpp
@@ -112,7 +112,7 @@ std::vector<TSharedRef> TChannelWriter::FlushBlock()
     return result;
 }
 
-int TChannelWriter::GetCurrentRowCount() const
+i64 TChannelWriter::GetCurrentRowCount() const
 {
     return CurrentRowCount;
 }
